cerberus_universal_backend/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.42
Flask-SQLAlchemy==3.1.1
cloud-sql-python-connector
python-dotenv==1.0.0
Flask-Migrate==4.0.5 # For database migrations
Flask-Bcrypt==1.0.1 # For password hashing
PyJWT==2.8.0 # For token-based authentication (optional, but good for APIs)
gunicorn==23.0.0 # WSGI server for deployment
Flask-CORS==6.0.1 # For Cross-Origin Resource Sharing
stripe==12.3.0
GeoAlchemy2
psycopg[binary]
google-cloud-secret-manager==2.16.2
google-cloud-secret-manager==2.16.2
